# perf(java): try to eliminate virtual function call for the kv serialization of Map<String, String>