<!--
**Thanks for contributing to Fury.**

**If this is your first time opening a PR on fury, you can refer to [CONTRIBUTING.md](https://github.com/apache/fury/blob/main/CONTRIBUTING.md).**

Contribution Checklist

    - The **Apache Fury (incubating)** community has restrictions on the naming of pr titles. You can also find instructions in [CONTRIBUTING.md](https://github.com/apache/fury/blob/main/CONTRIBUTING.md).

    - Fury has a strong focus on performance. If the PR you submit will have an impact on performance, please benchmark it first and provide the benchmark result here.
-->

## What does this PR do?

Add fast path for the serialization of `Map<String, String>`.

## Related issues

Closes #2026 

## Does this PR introduce any user-facing change?

None

## Benchmark

Inline log:
serialize:
```
@ 159   org.apache.fury.serializer.collection.AbstractMapSerializer::writeJavaChunk (318 bytes)   inline (hot)
  // ...
  @ 218   org.apache.fury.serializer.collection.AbstractMapSerializer::tryFastSerialize (33 bytes)   inline (hot)
    @ 20   org.apache.fury.serializer.StringSerializer::write (7 bytes)   inline (hot)
      @ 3   org.apache.fury.serializer.StringSerializer::writeJavaString (73 bytes)   inline (hot)
        @ 69   org.apache.fury.serializer.StringSerializer::writeCharsString (41 bytes)   inline (hot)
          @ 4   org.apache.fury.memory.Platform::getObject (9 bytes)   inline (hot)
            @ 5   sun.misc.Unsafe::getObject (0 bytes)   (intrinsic)
          @ 15   org.apache.fury.util.StringUtils::isLatin (6 bytes)   inline (hot)
            @ 2   org.apache.fury.util.StringUtils::isLatin (125 bytes)   inline (hot)
          @ 26   org.apache.fury.serializer.StringSerializer::writeCharsLatin1 (168 bytes)   inline (hot)
            @ 1   org.apache.fury.memory.MemoryBuffer::writerIndex (5 bytes)   accessor
            @ 21   org.apache.fury.memory.MemoryBuffer::ensure (14 bytes)   inline (hot)
            @ 25   org.apache.fury.memory.MemoryBuffer::getHeapMemory (5 bytes)   accessor
            @ 36   org.apache.fury.memory.MemoryBuffer::_unsafeHeapWriterIndex (10 bytes)   inline (hot)
            @ 53   org.apache.fury.memory.LittleEndian::putVarUint36Small (61 bytes)   inline (hot)
            @ 164   org.apache.fury.memory.MemoryBuffer::_unsafeWriterIndex (6 bytes)   inline (hot)
  @ 245   org.apache.fury.serializer.collection.AbstractMapSerializer::tryFastSerialize (33 bytes)   inline (hot)
    @ 20   org.apache.fury.serializer.StringSerializer::write (7 bytes)   inline (hot)
      @ 3   org.apache.fury.serializer.StringSerializer::writeJavaString (73 bytes)   inline (hot)
        @ 69   org.apache.fury.serializer.StringSerializer::writeCharsString (41 bytes)   inline (hot)
          @ 4   org.apache.fury.memory.Platform::getObject (9 bytes)   inline (hot)
            @ 5   sun.misc.Unsafe::getObject (0 bytes)   (intrinsic)
          @ 15   org.apache.fury.util.StringUtils::isLatin (6 bytes)   inline (hot)
            @ 2   org.apache.fury.util.StringUtils::isLatin (125 bytes)   inline (hot)
          @ 26   org.apache.fury.serializer.StringSerializer::writeCharsLatin1 (168 bytes)   inline (hot)
            @ 1   org.apache.fury.memory.MemoryBuffer::writerIndex (5 bytes)   accessor
            @ 21   org.apache.fury.memory.MemoryBuffer::ensure (14 bytes)   inline (hot)
            @ 25   org.apache.fury.memory.MemoryBuffer::getHeapMemory (5 bytes)   accessor
            @ 36   org.apache.fury.memory.MemoryBuffer::_unsafeHeapWriterIndex (10 bytes)   inline (hot)
            @ 53   org.apache.fury.memory.LittleEndian::putVarUint36Small (61 bytes)   inline (hot)
            @ 164   org.apache.fury.memory.MemoryBuffer::_unsafeWriterIndex (6 bytes)   inline (hot)
  // ...
```
deserialize:
```
@ 145   org.apache.fury.serializer.collection.AbstractMapSerializer::readJavaChunk (218 bytes)   inline (hot)
  // ...
  @ 143   org.apache.fury.serializer.collection.AbstractMapSerializer::tryFastDeserialize (24 bytes)   inline (hot)
    @ 14   org.apache.fury.serializer.StringSerializer::read (6 bytes)   inline (hot)
      @ 2   org.apache.fury.serializer.StringSerializer::readJavaString (64 bytes)   inline (hot)
        @ 60   org.apache.fury.serializer.StringSerializer::readCharsString (89 bytes)   inline (hot)
          @ 1   org.apache.fury.memory.MemoryBuffer::readVarUint36Small (89 bytes)   inline (hot)
            @ 21   org.apache.fury.memory.MemoryBuffer::_unsafeGetInt64 (33 bytes)   inlining too deep
            @ 85   org.apache.fury.memory.MemoryBuffer::readVarUint36Slow (122 bytes)   inlining too deep
          @ 29   org.apache.fury.serializer.StringSerializer::readCharsLatin1 (113 bytes)   inline (hot)
            @ 2   org.apache.fury.memory.MemoryBuffer::checkReadableBytes (29 bytes)   inlining too deep
            @ 6   org.apache.fury.memory.MemoryBuffer::getHeapMemory (5 bytes)   accessor
            @ 20   org.apache.fury.memory.MemoryBuffer::_unsafeHeapReaderIndex (10 bytes)   inlining too deep
            @ 59   org.apache.fury.memory.MemoryBuffer::_increaseReaderIndexUnsafe (11 bytes)   inlining too deep
          @ 85   org.apache.fury.serializer.StringSerializer::newCharsStringZeroCopy (32 bytes)   inline (hot)
            @ 23   java.lang.String$$Lambda$17/2003855659::apply (19 bytes)   inlining too deep
  @ 167   org.apache.fury.serializer.collection.AbstractMapSerializer::tryFastDeserialize (24 bytes)   inline (hot)
    @ 14   org.apache.fury.serializer.StringSerializer::read (6 bytes)   inline (hot)
      @ 2   org.apache.fury.serializer.StringSerializer::readJavaString (64 bytes)   inline (hot)
        @ 60   org.apache.fury.serializer.StringSerializer::readCharsString (89 bytes)   inline (hot)
          @ 1   org.apache.fury.memory.MemoryBuffer::readVarUint36Small (89 bytes)   inline (hot)
            @ 21   org.apache.fury.memory.MemoryBuffer::_unsafeGetInt64 (33 bytes)   inlining too deep
            @ 85   org.apache.fury.memory.MemoryBuffer::readVarUint36Slow (122 bytes)   inlining too deep
          @ 29   org.apache.fury.serializer.StringSerializer::readCharsLatin1 (113 bytes)   inline (hot)
            @ 2   org.apache.fury.memory.MemoryBuffer::checkReadableBytes (29 bytes)   inlining too deep
            @ 6   org.apache.fury.memory.MemoryBuffer::getHeapMemory (5 bytes)   accessor
            @ 20   org.apache.fury.memory.MemoryBuffer::_unsafeHeapReaderIndex (10 bytes)   inlining too deep
            @ 59   org.apache.fury.memory.MemoryBuffer::_increaseReaderIndexUnsafe (11 bytes)   inlining too deep
          @ 85   org.apache.fury.serializer.StringSerializer::newCharsStringZeroCopy (32 bytes)   inline (hot)
            @ 23   java.lang.String$$Lambda$17/2003855659::apply (19 bytes)   inlining too deep
  // ...
```