@jayhan94 2025-02-13
This PR isn't ready yet. I'll be adding a benchmark later. Maintainers, feel free to review it at the meantime, in case I've misunderstood the proposal discussed in #2026.