@jayhan94 2025-02-14
Hmm, I'm having trouble with the benchmark. I'm not sure whether the StringMapSerializationSuite is valid, but the results indicate that this PR introduces a significant performance regression.

```
Before:
Benchmark                                (mapSize)  (stringRatio)   Mode  Cnt        Score       Error  Units
StringMapSerializationSuite.deserialize         50            0.0  thrpt    5  1199813.656 ± 56659.770  ops/s
StringMapSerializationSuite.deserialize         50            0.2  thrpt    5  1082390.191 ± 22192.017  ops/s
StringMapSerializationSuite.deserialize         50            0.4  thrpt    5   988205.472 ± 10120.216  ops/s
StringMapSerializationSuite.deserialize         50            0.6  thrpt    5   819745.922 ± 14640.016  ops/s
StringMapSerializationSuite.deserialize         50            0.8  thrpt    5   761238.764 ± 41274.025  ops/s
StringMapSerializationSuite.deserialize         50            1.0  thrpt    5   818521.281 ± 42637.267  ops/s
StringMapSerializationSuite.serialize           50            0.0  thrpt    5  1313400.596 ± 41120.713  ops/s
StringMapSerializationSuite.serialize           50            0.2  thrpt    5  1197028.082 ± 38588.963  ops/s
StringMapSerializationSuite.serialize           50            0.4  thrpt    5  1226897.521 ± 28125.905  ops/s
StringMapSerializationSuite.serialize           50            0.6  thrpt    5  1137013.744 ± 54707.471  ops/s
StringMapSerializationSuite.serialize           50            0.8  thrpt    5  1094282.261 ± 12766.788  ops/s
StringMapSerializationSuite.serialize           50            1.0  thrpt    5  1330620.243 ± 37696.001  ops/s

After:
Benchmark                                (mapSize)  (stringRatio)   Mode  Cnt        Score       Error  Units
StringMapSerializationSuite.deserialize         50            0.0  thrpt    5  1013888.818 ± 24478.168  ops/s
StringMapSerializationSuite.deserialize         50            0.2  thrpt    5   784407.869 ± 40987.059  ops/s
StringMapSerializationSuite.deserialize         50            0.4  thrpt    5   762941.737 ± 36584.379  ops/s
StringMapSerializationSuite.deserialize         50            0.6  thrpt    5   875355.324 ±  7236.977  ops/s
StringMapSerializationSuite.deserialize         50            0.8  thrpt    5   765313.777 ± 16770.509  ops/s
StringMapSerializationSuite.deserialize         50            1.0  thrpt    5   854012.876 ±  8850.744  ops/s
StringMapSerializationSuite.serialize           50            0.0  thrpt    5  1377660.258 ± 13716.225  ops/s
StringMapSerializationSuite.serialize           50            0.2  thrpt    5  1250467.207 ±  6047.695  ops/s
StringMapSerializationSuite.serialize           50            0.4  thrpt    5  1162058.640 ±  5399.060  ops/s
StringMapSerializationSuite.serialize           50            0.6  thrpt    5  1213526.978 ±  4323.156  ops/s
StringMapSerializationSuite.serialize           50            0.8  thrpt    5  1021980.610 ± 28857.004  ops/s
StringMapSerializationSuite.serialize           50            1.0  thrpt    5  1280659.398 ±  2857.126  ops/s
```