@jayhan94 2025-02-18
> Hi @jayhan94 , do you try to mix the different kv type of map in one benchmark? For example, we have three map which have different type for key and value. In such cases, the jvm inline won't work, and the fastpath will kick in.

I mixed different types of key-value pairs(long,int,string) in single map, and I think the result is similar to having different types of maps, which is why I got the outcome above. The benchmark code is in the `StringMapSerializationSuite`.