@jayhan94 2025-02-18
> For example, we have three map which have different type for key and value.

I add a benchmark like this, and the result shows this pr has a performance improvement.
```
Before:
Benchmark                                 (mapSize)   Mode  Cnt       Score       Error  Units
StringMapSerializationSuite.deserialize2         50  thrpt    5  368898.317 ± 21006.990  ops/s
StringMapSerializationSuite.serialize2           50  thrpt    5  590228.934 ± 21081.928  ops/s

After:
Benchmark                                 (mapSize)   Mode  Cnt       Score      Error  Units
StringMapSerializationSuite.deserialize2         50  thrpt    5  372951.992 ± 5156.001  ops/s
StringMapSerializationSuite.serialize2           50  thrpt    5  748249.524 ± 9676.674  ops/s
```